+++ AudioKit/Common/Nodes/Generators/Oscillators/AKSquareWaveOscillator/AKSquareWaveOscillator.swift
@@ -23,21 +23,21 @@ public struct AKSquareWaveOscillator: AKNode {
     private var pulseWidthParameter: AUParameter?
 
     /** In cycles per second, or Hz. */
-    public var frequency: Float = 440 {
+    public var frequency: Double = 440 {
         didSet {
-            frequencyParameter?.setValue(frequency, originator: token!)
+            frequencyParameter?.setValue(Float(frequency), originator: token!)
         }
     }
     /** Output amplitude */
-    public var amplitude: Float = 1.0 {
+    public var amplitude: Double = 1.0 {
         didSet {
-            amplitudeParameter?.setValue(amplitude, originator: token!)
+            amplitudeParameter?.setValue(Float(amplitude), originator: token!)
         }
     }
     /** Duty cycle width (range 0-1). */
-    public var pulseWidth: Float = 0.5 {
+    public var pulseWidth: Double = 0.5 {
         didSet {
-            pulseWidthParameter?.setValue(pulseWidth, originator: token!)
+            pulseWidthParameter?.setValue(Float(pulseWidth), originator: token!)
         }
     }
 
@@ -45,9 +45,9 @@ public struct AKSquareWaveOscillator: AKNode {
 
     /** Initialize this oscillator node */
     public init(
-        frequency: Float = 440,
-        amplitude: Float = 1.0,
-        pulseWidth: Float = 0.5) {
+        frequency: Double = 440,
+        amplitude: Double = 1.0,
+        pulseWidth: Double = 0.5) {
 
         self.frequency = frequency
         self.amplitude = amplitude
@@ -88,16 +88,16 @@ public struct AKSquareWaveOscillator: AKNode {
 
             dispatch_async(dispatch_get_main_queue()) {
                 if address == self.frequencyParameter!.address {
-                    self.frequency = value
+                    self.frequency = Double(value)
                 } else if address == self.amplitudeParameter!.address {
-                    self.amplitude = value
+                    self.amplitude = Double(value)
                 } else if address == self.pulseWidthParameter!.address {
-                    self.pulseWidth = value
+                    self.pulseWidth = Double(value)
                 }
             }
         }
-        frequencyParameter?.setValue(frequency, originator: token!)
-        amplitudeParameter?.setValue(amplitude, originator: token!)
-        pulseWidthParameter?.setValue(pulseWidth, originator: token!)
+        frequencyParameter?.setValue(Float(frequency), originator: token!)
+        amplitudeParameter?.setValue(Float(amplitude), originator: token!)
+        pulseWidthParameter?.setValue(Float(pulseWidth), originator: token!)
     }
 }
